refactor(ui): break up the internal transfer confirm sheet
The file was 1083 lines and held four types, three of them already used from
other screens — `HomeView`, `CoinJoinMoveFundsSheet` and `SendScreen` were all
reaching into the confirm sheet's file for a step list, a terminal state view
and a recovery sheet. Those move out, and the shielded machinery they belong
with (`ShieldedTransferCoordinator`, `ShieldedWithdrawalStore`) joins them in
`Shielded/`.

`InternalTransferSummaryFigures` takes the numbers. Pricing four of the six
routes through the SDK and reconstructing Core → Shielded's executed lock value
is fee math, which `CLAUDE.md` keeps out of a `View` — and it is also why none
of it could be exercised without rendering a sheet. Each figure now returns nil
when it cannot be computed and the sheet renders the em dash, so failing closed
and drawing a dash stopped being one decision.

`TransferPrivacyTip` takes the route-to-copy table, redrawn on
`SystemMessageView` instead of a hand-built circle-and-two-labels card. It read
only the route and the full-withdrawal flag, so all six routes are previewable
on their own now.

The sheet itself moves onto `DashUIKit.BottomSheet` — grabber, title and
background come from the design system rather than being drawn inline.

551 lines left, of which 78 are previews.
